02/16 17:00 ユーザー情報編集画面の修正 & init.sqlのicon_pathカラムにUNITQUE追加 butch

+++ ChatApp/app.py
@@ -110,24 +110,26 @@ def edit_user():
         is_instructor = request.form.get('is_instructor')
         user = dbConnect.getUserById(u_id)
         icon_name = secure_filename(file.filename)
-        
         # 新しい画像ファイルを受取ったか確認
         if icon_name == '':
             icon_path = user['icon_path']
         else:
+            # 同名のファイルの有無をチェック
+            icon_path = os.path.join('static/img_add', icon_name)
+            icon_paths = dbConnect.getIconPath()
+            if icon_path in icon_paths:
+                flash('画像のファイル名を変更してください')
+                return redirect('/edit-user')
             # 古い画像ファイルの削除
             if user['icon_path'] is not None:
                 os.remove(user['icon_path'])
             # 受取った写真（icon）をimg_addフォルダに保存
-            icon_path = os.path.join('static/img_add', icon_name)
             file.save(icon_path)
-        
         # 新しいpasswordを受取ったか確認
         if password == '':
             hash_password = user['password']
         else:
             hash_password = hashlib.sha256(password.encode('utf-8')).hexdigest()
-        
         dbConnect.updateUser(u_id, user_name, mail, hash_password, is_instructor, height, goal, introduction, address, icon_path)
         # マイページができたら/mypageへリダイレクト
         return redirect('/edit-user')
